+++ tap-snapshots/lib-test-TAP.test.js
@@ -17,7 +17,7 @@ exports[`lib/test TAP graphqlFetchOptions with a standard operation. > Fetch opt
 }
 `
 
-exports[`lib/test TAP Query SSR with fetch unavailable. > Console log. 1`] = `
+exports[`lib/test TAP SSR Query with fetch unavailable. > Console log. 1`] = `
 [
   "GraphQL request (hash “[hash]”) errors:",
   "  Fetch:",
@@ -25,13 +25,13 @@ exports[`lib/test TAP Query SSR with fetch unavailable. > Console log. 1`] = `
 ]
 `
 
-exports[`lib/test TAP Query SSR with fetch unavailable. > GraphQL request cache. 1`] = `
+exports[`lib/test TAP SSR Query with fetch unavailable. > GraphQL request cache. 1`] = `
 {
   "fetchError": "Global fetch API or polyfill unavailable."
 }
 `
 
-exports[`lib/test TAP Query SSR with fetch unavailable. > Query render function arguments. 1`] = `
+exports[`lib/test TAP SSR Query with fetch unavailable. > Query render function arguments. 1`] = `
 {
   "0": {
     "load": "[Function]",
@@ -41,7 +41,7 @@ exports[`lib/test TAP Query SSR with fetch unavailable. > Query render function 
 }
 `
 
-exports[`lib/test TAP Query SSR with relative fetch URL. > Console log. 1`] = `
+exports[`lib/test TAP SSR Query with relative fetch URL. > Console log. 1`] = `
 [
   "GraphQL request (hash “[hash]”) errors:",
   "  Fetch:",
@@ -49,13 +49,13 @@ exports[`lib/test TAP Query SSR with relative fetch URL. > Console log. 1`] = `
 ]
 `
 
-exports[`lib/test TAP Query SSR with relative fetch URL. > GraphQL request cache. 1`] = `
+exports[`lib/test TAP SSR Query with relative fetch URL. > GraphQL request cache. 1`] = `
 {
   "fetchError": "Only absolute URLs are supported"
 }
 `
 
-exports[`lib/test TAP Query SSR with relative fetch URL. > Query render function arguments. 1`] = `
+exports[`lib/test TAP SSR Query with relative fetch URL. > Query render function arguments. 1`] = `
 {
   "0": {
     "load": "[Function]",
@@ -65,7 +65,7 @@ exports[`lib/test TAP Query SSR with relative fetch URL. > Query render function
 }
 `
 
-exports[`lib/test TAP Query SSR with HTTP error. > Console log. 1`] = `
+exports[`lib/test TAP SSR Query with HTTP error. > Console log. 1`] = `
 [
   "GraphQL request (hash “[hash]”) errors:",
   "  HTTP:",
@@ -76,7 +76,7 @@ exports[`lib/test TAP Query SSR with HTTP error. > Console log. 1`] = `
 ]
 `
 
-exports[`lib/test TAP Query SSR with HTTP error. > GraphQL request cache. 1`] = `
+exports[`lib/test TAP SSR Query with HTTP error. > GraphQL request cache. 1`] = `
 {
   "httpError": {
     "status": 404,
@@ -86,7 +86,7 @@ exports[`lib/test TAP Query SSR with HTTP error. > GraphQL request cache. 1`] = 
 }
 `
 
-exports[`lib/test TAP Query SSR with HTTP error. > Query render function arguments. 1`] = `
+exports[`lib/test TAP SSR Query with HTTP error. > Query render function arguments. 1`] = `
 {
   "0": {
     "load": "[Function]",
@@ -100,7 +100,7 @@ exports[`lib/test TAP Query SSR with HTTP error. > Query render function argumen
 }
 `
 
-exports[`lib/test TAP Query SSR with response JSON invalid. > Console log. 1`] = `
+exports[`lib/test TAP SSR Query with response JSON invalid. > Console log. 1`] = `
 [
   "GraphQL request (hash “[hash]”) errors:",
   "  Parse:",
@@ -108,13 +108,13 @@ exports[`lib/test TAP Query SSR with response JSON invalid. > Console log. 1`] =
 ]
 `
 
-exports[`lib/test TAP Query SSR with response JSON invalid. > GraphQL request cache. 1`] = `
+exports[`lib/test TAP SSR Query with response JSON invalid. > GraphQL request cache. 1`] = `
 {
   "parseError": "invalid json response body at http://localhost:<port>/ reason: Unexpected token N in JSON at position 0"
 }
 `
 
-exports[`lib/test TAP Query SSR with response JSON invalid. > Query render function arguments. 1`] = `
+exports[`lib/test TAP SSR Query with response JSON invalid. > Query render function arguments. 1`] = `
 {
   "0": {
     "load": "[Function]",
@@ -124,7 +124,7 @@ exports[`lib/test TAP Query SSR with response JSON invalid. > Query render funct
 }
 `
 
-exports[`lib/test TAP Query SSR with response payload malformed. > Console log. 1`] = `
+exports[`lib/test TAP SSR Query with response payload malformed. > Console log. 1`] = `
 [
   "GraphQL request (hash “[hash]”) errors:",
   "  Parse:",
@@ -132,13 +132,13 @@ exports[`lib/test TAP Query SSR with response payload malformed. > Console log. 
 ]
 `
 
-exports[`lib/test TAP Query SSR with response payload malformed. > GraphQL request cache. 1`] = `
+exports[`lib/test TAP SSR Query with response payload malformed. > GraphQL request cache. 1`] = `
 {
   "parseError": "Malformed payload."
 }
 `
 
-exports[`lib/test TAP Query SSR with response payload malformed. > Query render function arguments. 1`] = `
+exports[`lib/test TAP SSR Query with response payload malformed. > Query render function arguments. 1`] = `
 {
   "0": {
     "load": "[Function]",
@@ -148,7 +148,7 @@ exports[`lib/test TAP Query SSR with response payload malformed. > Query render 
 }
 `
 
-exports[`lib/test TAP Query SSR with GraphQL errors. > Console log. 1`] = `
+exports[`lib/test TAP SSR Query with GraphQL errors. > Console log. 1`] = `
 [
   "GraphQL request (hash “[hash]”) errors:",
   "  HTTP:",
@@ -159,7 +159,7 @@ exports[`lib/test TAP Query SSR with GraphQL errors. > Console log. 1`] = `
 ]
 `
 
-exports[`lib/test TAP Query SSR with GraphQL errors. > GraphQL request cache. 1`] = `
+exports[`lib/test TAP SSR Query with GraphQL errors. > GraphQL request cache. 1`] = `
 {
   "httpError": {
     "status": 400,
@@ -179,7 +179,7 @@ exports[`lib/test TAP Query SSR with GraphQL errors. > GraphQL request cache. 1`
 }
 `
 
-exports[`lib/test TAP Query SSR with GraphQL errors. > Query render function arguments. 1`] = `
+exports[`lib/test TAP SSR Query with GraphQL errors. > Query render function arguments. 1`] = `
 {
   "0": {
     "load": "[Function]",
@@ -203,7 +203,7 @@ exports[`lib/test TAP Query SSR with GraphQL errors. > Query render function arg
 }
 `
 
-exports[`lib/test TAP Query SSR with variables. > GraphQL request cache. 1`] = `
+exports[`lib/test TAP SSR Query with variables. > GraphQL request cache. 1`] = `
 {
   "data": {
     "date": {
@@ -213,7 +213,7 @@ exports[`lib/test TAP Query SSR with variables. > GraphQL request cache. 1`] = `
 }
 `
 
-exports[`lib/test TAP Query SSR with variables. > Query render function arguments. 1`] = `
+exports[`lib/test TAP SSR Query with variables. > Query render function arguments. 1`] = `
 {
   "0": {
     "load": "[Function]",
@@ -227,7 +227,7 @@ exports[`lib/test TAP Query SSR with variables. > Query render function argument
 }
 `
 
-exports[`lib/test TAP Query SSR with nested query. > HTML displaying the nested query render function argument. 1`] = `
+exports[`lib/test TAP SSR Query with nested query. > HTML displaying the nested query render function argument. 1`] = `
 <pre>{
   "load": "[Function]",
   "loading": true,
@@ -237,6 +237,6 @@ exports[`lib/test TAP Query SSR with nested query. > HTML displaying the nested 
 }</pre>
 `
 
-exports[`lib/test TAP Preload legacy React context API components. > HTML. 1`] = `
+exports[`lib/test TAP SSR legacy React context API components. > HTML. 1`] = `
 <div data-reactroot=""><div><p>Context value.</p></div></div>
 `