+++ changelog.md
@@ -2,10 +2,21 @@
 
 ## Next
 
+### Breaking
+
+- Removed the `preload` function. It was not going to work with React hooks.
+- Added the [`react-dom`](https://npm.im/react-dom) peer dependency.
+
+### Minor
+
+- Added the `ssr` function, which is for server use only and is React hooks ready. It is simpler and more future-proof than the removed `preload` function as it leverages [`ReactDOMServer`](https://reactjs.org/docs/react-dom-server) for rendering.
+- `GraphQL` now emits a `cache` promise in the `fetch` event payload. These events are undocumented, so this could be considered an internal change.
+
 ### Patch
 
 - Updated dependencies.
 - Handle exceptions outside tests (see [tapjs/node-tap#463 (comment)](https://github.com/tapjs/node-tap/issues/463#issuecomment-456701261)).
+- Added a `ReactNode` JSDoc type, replacing `ReactElement` types.
 
 ## 6.0.1